refactor to use tabs design
